user options

+++ infiniswap_daemon/rdma-common.c
@@ -422,7 +422,7 @@ void* free_mem(void *data)
 
   rdma_session_init(&session);
   last_free_mem_g = (int)(get_free_mem() / ONE_MB);
-  printf("%s, is called, last %d GB, weight: %f, %f\n", __func__, last_free_mem_g, CURR_FREE_MEM_WEIGHT, last_free_mem_weight); 
+  printf("%s, is called, last %d GB, weight: %f, %f\n", __func__, last_free_mem_g, (float)(CURR_FREE_MEM_WEIGHT), last_free_mem_weight); 
 
   while (running) {// server is working
     free_mem_g = (int)(get_free_mem() / ONE_MB);
